Split file into components
