+++ src/devtools/index.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
-import { Ships } from "./Ships";
+import { Ships } from "./ships";
 import "@alenaksu/json-viewer";
 
 export const DevTools: React.FC = () => (
